+++ panweave-smart-energy/src/clusters/mod.rs
@@ -4,3 +4,4 @@ pub mod drlc;
 pub mod messaging;
 pub mod metering;
 pub mod price;
+pub mod tunneling;